Fix CLANG build

+++ util/hash.h
@@ -39,7 +39,8 @@ inline uint32_t GetSliceHash(const Slice& s) {
 inline uint64_t NPHash64(const char* data, size_t n, uint32_t seed) {
   // Similar to murmur hash, but reduce one multiplication
   // out of two. The hash quality is not as good as murmurhash.
-  return MURMUR_HASH(data, n, static_cast<unsigned int>(seed));
+  return MURMUR_HASH(data, static_cast<int>(n),
+                     static_cast<unsigned int>(seed));
 }
 
 // std::hash compatible interface.
